Added upgradeVillage() and calculateCost()
In VillageComponent.cs, upgradeVillage() and calculateCost()
were implemented as per the upgradeVillage sequence diagram.

Still unsure of calculateCost - an upgradeCosts array proved
to be overly complex for what we're trying to do - used switch
statements instead to calculate the cost.

TODO: Implement a way of comparing enums for VillageComponent and
UnitComponent in the upgrade methods. Currently comparing with
basic comparison which isn't really valid for enums.